Created msgs package
Added new msg package and cleaned up the  CMakeLists
