+++ gazebo_model_attachment_plugin/.gitignore
