+++ gazebo_model_attachment_plugin/.rosinstall
