+++ gazebo_model_attachment_plugin/setup.py
