+++ gazebo_model_attachment_plugin_msgs/.gitignore
@@ -1,0 +1,25 @@
+.idea
+
+build
+devel
+install
+public
+
+*.pyc
+*~
+.user
+.catkin_tools/
+.autosave
+logs/
+*.user
+
+*.docker
+*.codelite
+*.project
+*.workspace
+*.tags
+.codelite/
+.history
+.vscode
+
+.mypy_cache/
