+++ gazebo_model_attachment_plugin_msgs/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="3">
-    <name>gazebo_model_attachment_plugin</name>
+    <name>gazebo_model_attachment_plugin_msgs</name>
     <version>1.0.0</version>
     <description>Model Attachment Plugin</description>
     <maintainer email="Jason.Cochrane@boeing.com">Boeing</maintainer>
@@ -12,9 +12,6 @@
     <buildtool_depend>ament_cmake</buildtool_depend>
     <buildtool_depend>ament_cmake_python</buildtool_depend>
 
-    <depend>gazebo</depend>
-    <depend>gazebo_msgs</depend>
-    <depend>gazebo_ros</depend>
     <depend>rclcpp</depend>
     <depend>std_msgs</depend>
 
@@ -29,6 +26,6 @@
 
     <export>
         <build_type>ament_cmake</build_type>
-        <gazebo_ros plugin_path="${prefix}/lib" gazebo_media_path="${prefix}"/>
+        <gazebo_ros plugin_path="${prefix}/lib" gazebo_media_path="${prefix}" />
     </export>
 </package>
