+++ gazebo_model_attachment_plugin_msgs/srv/Attach.srv
